refactor(mps-sync-plugin-lib): refactoring when state is perssisted and restored
1. Move PersistableState to library and rename Configuration class to SyncPluginState. 2. After state restore, set the active connection and branch according to the restored state.
