+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/ModelSyncService.kt
@@ -46,6 +46,7 @@ import org.modelix.model.lazy.BranchReference
 import org.modelix.model.lazy.CLVersion
 import org.modelix.model.lazy.RepositoryId
 import org.modelix.mps.sync.IBinding
+import org.modelix.mps.sync.IRebindSyncService
 import org.modelix.mps.sync.ISyncService
 import org.modelix.mps.sync.SyncServiceImpl
 import org.modelix.mps.sync.mps.notifications.BalloonNotifier
@@ -56,12 +57,11 @@ import java.net.URL
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
 @Service(Service.Level.APP)
-class ModelSyncService : Disposable {
+class ModelSyncService : Disposable, IRebindSyncService {
 
     private val logger = KotlinLogging.logger { }
     private val notifierInjector = InjectableNotifierWrapper
 
-    private val unattendedClients = mutableListOf<ModelClientV2>()
     private lateinit var syncService: ISyncService
 
     init {
@@ -70,16 +70,31 @@ class ModelSyncService : Disposable {
         logger.debug { "ModelixSyncPlugin: Registration finished" }
     }
 
-    fun connectModelServer(url: String, jwt: String): ModelClientV2? {
+    override fun connectModelServer(serverURL: String, jwt: String?): ModelClientV2? {
         var client: ModelClientV2? = null
         try {
-            client = syncService.connectModelServer(URL(url), jwt)
-            notifierInjector.notifyAndLogInfo("Connected to server: $url", logger)
+            client = syncService.connectModelServer(URL(serverURL), jwt)
+            notifierInjector.notifyAndLogInfo("Connected to server: $serverURL", logger)
         } catch (t: Throwable) {
-            val message = "Unable to connect to $url. Cause: ${t.message}"
+            val message = "Unable to connect to $serverURL. Cause: ${t.message}"
             notifierInjector.notifyAndLogError(message, t, logger)
         }
         return client
+    }
+
+    override fun rebindModules(
+        client: ModelClientV2,
+        branchReference: BranchReference,
+        initialVersion: CLVersion,
+        modules: Iterable<AbstractModule>,
+    ): Iterable<IBinding>? {
+        try {
+            return syncService.rebindModules(client, branchReference, initialVersion, modules)
+        } catch (t: Throwable) {
+            val message = "Error while binding modules to Branch '$branchReference'. Cause: ${t.message}"
+            notifierInjector.notifyAndLogError(message, t, logger)
+            return null
+        }
     }
 
     fun disconnectServer(modelClient: ModelClientV2): ModelClientV2? {
@@ -139,21 +154,6 @@ class ModelSyncService : Disposable {
         }
     }
 
-    fun rebindModules(
-        client: ModelClientV2,
-        branchReference: BranchReference,
-        initialVersion: CLVersion,
-        modules: Iterable<AbstractModule>,
-    ): Iterable<IBinding>? {
-        try {
-            return syncService.rebindModules(client, branchReference, initialVersion, modules)
-        } catch (t: Throwable) {
-            val message = "Error while binding modules to Branch '$branchReference'. Cause: ${t.message}"
-            notifierInjector.notifyAndLogError(message, t, logger)
-            return null
-        }
-    }
-
     fun bindModuleFromMps(module: AbstractModule, branch: IBranch) = syncService.bindModuleFromMps(module, branch)
 
     fun bindModelFromMps(model: SModelBase, branch: IBranch) = syncService.bindModelFromMps(model, branch)
@@ -172,19 +172,9 @@ class ModelSyncService : Disposable {
         logger.debug { "ModelixSyncPlugin: EnsureStarted" }
     }
 
-    fun registerUnattendedClient(client: ModelClientV2) {
-        unattendedClients.add(client)
-    }
-
     override fun dispose() {
         logger.debug { "ModelixSyncPlugin: Dispose" }
-        try {
-            unattendedClients.forEach(::disconnectServer)
-        } catch (_: Throwable) {
-            // ignored
-        } finally {
-            syncService.close()
-        }
+        syncService.close()
     }
 
     private fun registerSyncActions() {
