+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/configuration/SyncPluginState.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright (c) 2023-2024.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.modelix.mps.sync.plugin.configuration
+
+import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.RoamingType
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.State
+import com.intellij.openapi.components.Storage
+import com.intellij.openapi.components.service
+import org.modelix.kotlin.utils.UnstableModelixFeature
+import org.modelix.mps.sync.persistence.PersistableState
+import org.modelix.mps.sync.plugin.ModelSyncService
+
+@UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
+@Service(Service.Level.PROJECT)
+@State(
+    name = "ModelixSyncPluginState",
+    reloadable = true,
+    storages = [Storage("modelixSyncPluginState.xml", roamingType = RoamingType.DISABLED)],
+)
+class SyncPluginState : PersistentStateComponent<PersistableState> {
+
+    var latestLoadedState: PersistableState? = null
+        private set
+
+    override fun getState(): PersistableState {
+        return PersistableState().getCurrentState()
+    }
+
+    override fun loadState(newState: PersistableState) {
+        val syncService = service<ModelSyncService>()
+        newState.load(syncService)
+        latestLoadedState = newState
+    }
+}
