+++ mps-sync-plugin/src/main/resources/META-INF/plugin.xml
@@ -38,9 +38,9 @@
                     icon="org.modelix.mps.sync.plugin.icons.CloudIcons.PLUGIN_ICON" anchor="right"
                     factoryClass="org.modelix.mps.sync.plugin.gui.ModelSyncGuiFactory"/>
 
-        <!-- service to handle storage of cloud details via project settings, aka .mps/modelix.xml -->
+        <!-- service to handle storage of cloud details via project settings, aka .mps/modelixSyncPluginState.xml -->
         <projectService
-                serviceImplementation="org.modelix.mps.sync.plugin.configuration.CloudResourcesConfigurationComponent"/>
+                serviceImplementation="org.modelix.mps.sync.plugin.configuration.SyncPluginState"/>
 
         <!-- the actual service in charge of connecting and binding -->
         <applicationService serviceImplementation="org.modelix.mps.sync.plugin.ModelSyncService"/>
